feat(training): add A7 callbacks + dummy test modules (Phase 4a)
NaNLossGuard / GradScalerWatchdog / EventEmitter / RngStateCallback per
A7, plus the test-only _DummyBackbone/_DummyHead/_LossReturningScalar
make_test_module building blocks (A14). pyrightconfig gains
tests-root extraPaths so the cross-test-package import resolves; tidies
an int(round()) in sampling.py.

+++ src/seq_sklearn/training/sampling.py
@@ -56,7 +56,7 @@ def oversample_minority(
     """
     by_class = _class_indices(labels)
     majority = max(len(idx) for idx in by_class.values())
-    target = int(round(oversample_ratio * majority))
+    target = round(oversample_ratio * majority)
 
     pieces: list[np.ndarray] = []
     for _cls, idx in sorted(by_class.items()):
